remove session concern
