+++ lib/shopify_app.rb
@@ -17,7 +17,6 @@ require 'shopify_app/scripttags_manager_job'
 # helpers and concerns
 require 'shopify_app/shop'
 require 'shopify_app/session_storage'
-require 'shopify_app/sessions_concern'
 require 'shopify_app/localization'
 require 'shopify_app/login_protection'
 require 'shopify_app/webhooks_manager'
